x86: save/restore 64-bit values of XMM registers if CPU supports it

+++ arch/x86/include/arch/stack-frame.h
@@ -18,7 +18,7 @@ struct native_stack_frame {
 
 struct jit_stack_frame {
 	void *prev; /* previous stack frame link */
-	unsigned long xmm[8];
+	unsigned long long xmm[8];
 	unsigned long old_ebx;
 	unsigned long old_esi;
 	unsigned long old_edi;
